requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
requests
pyautogui
mss
pillow
pydantic
google-cloud-firestore
